requirements.txt
python-telegram-bot>=20.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
